let basic calculator works
